refactor: split load_layer_map in two
